+++ src/main/java/com/ordertogether/team14_be/order/details/dto/get/SmsReq.java
@@ -1,3 +1,3 @@
-package com.ordertogether.team14_be.order.details.sms.dto;
+package com.ordertogether.team14_be.order.details.dto.get;
 
 public record SmsReq(Long spotId) {}
